Version v3 of the script 'variabledNLsimulation_v3.py'
Script variabledNLsimulation_v3.py is used to analyze the transition of a beam in an Iota ring through resonance during variation of the strength of nonlinear lenses. Folder pictures_v3_07172019 contains plots to describe the transition of a beam in an Iota ring through the resonance when t=0.5 in 4 different cases: final strengths of central nonlinear lens are 0.5, 0.8, 1.0, and 1.2 correspondingly. Еransition is carried out within 200 turns: the strengths of the lenses increases from initial value (t=0.0 for central lens)  to the final values in steps every 10 turns.